requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==1.0.3
pyparsing==3.1.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.4
WTForms==3.0.1
